doc/examples: add upper bound to minimal CMake version
Add an upper bound to the minimal supported CMake version. This keeps
compatibility with CMake 3.5 while also telling newer versions that the
syntax was tested against this upper bound. The 3.31.0 version was
chosen as it this the newest one I could test at the moment.

This fixes the following warning when building with CMake 3.31:

 CMake Deprecation Warning at CMakeLists.txt:5 (cmake_minimum_required):
  Compatibility with CMake < 3.10 will be removed from a future version of
  CMake.

  Update the VERSION argument <min> value.  Or, use the <min>...<max> syntax
  to tell CMake that the project requires at least <min> but has been updated
  to work with policies introduced by <max> or earlier.

Change-Id: Iffeb9326341e11aba500cf519bf6e131b64dd9e8
Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>